validation du tp5

+++ mathieu-pizzeria-project/mathieu-pizzeria-parent/mathieu-pizzeria-dao/src/main/java/fr/pizzeria/dao/pizza/IPizzaRepository.java
@@ -14,5 +14,5 @@ import fr.pizzeria.model.Pizza;
  * @author mathieu
  *
  */
-public interface IPizzaRepository  extends JpaRepository<Pizza, Long>  {
+public interface IPizzaRepository  extends JpaRepository<Pizza, Integer>  {
 }
